installer improvements

+++ Scope.Installer/Installer.cs
@@ -8,11 +8,16 @@
 namespace Scope.Installer
 {
     using System;
+    using System.Collections.Generic;
     using System.IO;
     using System.IO.Compression;
     using System.Linq;
+    using System.Net;
     using System.Net.Http;
+    using System.Security.Cryptography;
+    using System.Text.RegularExpressions;
     using System.Threading.Tasks;
+    using Microsoft.Win32;
 
     /// <summary>
     /// The installer tool.
@@ -23,13 +28,16 @@ namespace Scope.Installer
         /// The URL to download the Scope files from.
         /// </summary>
         private const string DownloadUrl = "https://scopesl.com/ScopeStuff.zip";
-
-        // TODO: Autodetect game path.
-
+        
         /// <summary>
-        /// The game folder path.
+        /// The download hash.
         /// </summary>
-        private const string GameFolder = @"C:\Program Files (x86)\Steam\steamapps\common\SCP Secret Laboratory";
+        private const string ZipHash = "DD567472AFF7FCF14916EF3AF4E80EA58BB02AF8200EEF9BED639307AF5A2D67";
+
+        private static readonly List<string> DoNotOverwrite = new()
+        {
+            "SCPSL.exe",
+        };
 
         /// <summary>
         /// This method is the entrypoint of the program.
@@ -37,9 +45,26 @@ namespace Scope.Installer
         /// <param name="args"><see cref="string"/>[] args.</param>
         public static async Task Main(string[] args)
         {
+            string GameFolder = GetSLFolder();
+
             try
             {
                 var download = await Download(DownloadUrl);
+                var sha256 = SHA256.Create();
+                var hash = BitConverter.ToString(sha256.ComputeHash(download)).Replace("-", string.Empty);
+
+                using (WebClient wc = new WebClient())
+                {
+                    wc.DownloadFile("https://github.com/moddedmcplayer/scope-files/raw/main/ScopeStuff.zip", "ScopeStuff.zip");
+                }
+
+                if (hash != ZipHash)
+                {
+                    Console.WriteLine("The archive hash does not match!");
+                    Console.Read();
+                    Environment.Exit(0);
+                }
+
                 var archive = new ZipArchive(download);
                 if (archive.Entries.All(x => !x.FullName.StartsWith("ScopeStuff")))
                 {
@@ -54,21 +79,36 @@ namespace Scope.Installer
                     return;
                 }
 
+                var AC = Path.Combine(GameFolder, "SL-AC.dll");
+                if (File.Exists(AC))
+                {
+                    File.Move(AC, $"{AC}.disabled");
+                }
+
                 foreach (var file in Directory.GetFiles(Path.Combine(GameFolder, "ScopeStuff")))
                 {
-                    var originalPath = Path.Combine(GameFolder, file
+                    var fileName = file
                         .Substring(file.Length - file
                             .ToCharArray()
                             .Reverse()
                             .ToList()
-                            .IndexOf(Path.DirectorySeparatorChar)));
+                            .IndexOf(Path.DirectorySeparatorChar));
+                    var originalPath = Path.Combine(GameFolder, fileName);
 
-                    if (File.Exists(originalPath))
+                    if (File.Exists(originalPath) && DoNotOverwrite.Contains(fileName))
                     {
-                        File.Move(originalPath, $"{originalPath}.old", true);
+                        if (!File.Exists($"{originalPath}.old"))
+                        {
+                            File.Move(originalPath, $"{originalPath}.old", true);
+                        }
                     }
 
                     File.Move(file, originalPath, true);
+                }
+
+                foreach (var dir in Directory.GetDirectories(Path.Combine(GameFolder, "ScopeStuff")))
+                {
+                    Directory.Move(dir, dir.Replace("ScopeStuff\\", string.Empty));
                 }
             }
             catch (Exception ex)
@@ -92,6 +132,7 @@ namespace Scope.Installer
             {
                 Timeout = TimeSpan.FromSeconds(480),
             };
+            client.DefaultRequestHeaders.Add("User-Agent", "Scope Installer");
 
             try
             {
@@ -111,6 +152,73 @@ namespace Scope.Installer
             }
 
             return stream;
+        }
+
+        private const string DefaultSteamPath = @"C:\Program Files (x86)\Steam";
+        private const string DefaultSLPath = @"C:\Program Files (x86)\Steam\steamapps\common\SCP Secret Laboratory";
+        private const string SLFolderName = @"SCP Secret Laboratory";
+
+        private static string GetSLFolder()
+        {
+            string steamPath;
+            RegistryKey softwareKey = null;
+            try
+            {
+                string softwarePath = Environment.Is64BitProcess ? @"SOFTWARE\WOW6432Node" : "SOFTWARE";
+                softwareKey = Registry.LocalMachine.OpenSubKey(softwarePath, false);
+
+                using (var key = softwareKey.OpenSubKey(@"Valve\Steam"))
+                {
+                    var obj = key.GetValue("InstallPath");
+                    var installPath = obj as string;
+                    bool foundSteam = (!string.IsNullOrWhiteSpace(installPath) && Directory.Exists(installPath));
+
+                    steamPath = foundSteam ? installPath : DefaultSteamPath;
+                }
+            }
+            catch
+            {
+                steamPath = DefaultSteamPath;
+            }
+            finally
+            {
+                if (softwareKey != null)
+                    softwareKey.Close();
+            }
+
+            string SLPath = DefaultSLPath;
+            var libraryPaths = new List<string>();
+
+            var vdfFile = new FileInfo(Path.Combine(steamPath, @"steamapps\libraryfolders.vdf"));
+            if (!vdfFile.Exists)
+            {
+                string content;
+                using (var stream = vdfFile.OpenRead())
+                {
+                    using (var reader = new StreamReader(stream))
+                    {
+                        content = reader.ReadToEnd();
+                    }
+                }
+
+                var matches = Regex.Matches(content, "\"\\d\"\\s+\"(?<path>.+)\"");
+                foreach (Match match in matches)
+                {
+                    string path = match.Groups["path"].Value;
+                    path = path.Replace(@"\\", @"\");
+                    libraryPaths.Add(path);
+                }
+
+                foreach (var path in libraryPaths)
+                {
+                    if (Directory.Exists(path) && Directory.GetDirectories(path).Contains(SLFolderName))
+                    {
+                        SLPath = path;
+                    }
+                }
+            }
+
+            return SLPath;
         }
     }
 }
